+++ com.sparta.logistics.delivery-service/src/main/java/com/sparta/logistics/delivery_service/application/service/ProducerService.java
@@ -1,6 +1,6 @@
 package com.sparta.logistics.delivery_service.application.service;
 
-import com.sparta.logistics.delivery_service.application.dto.DeliveryInfoDto;
+import com.sparta.logistics.delivery_service.infrastructure.messaging.dto.DeliveryInfoDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
